% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sazed.R
\name{azed}
\alias{azed}
\title{Compute the AZED component of the SAZED ensemble}
\usage{
azed(y, preprocess = T)
}
\arguments{
\item{y}{The input time series.}

\item{preprocess}{If true, y is detrended and z-normalized before
computation.}
}
\value{
The AZED season length estimate of y.
}
\description{
\code{azed} computes the autocorrelation of its argument, and then derives the
season length from its the autocorrelations zero density.
}
\examples{
season_length <- 26
y <- sin(1:400*2*pi/season_length)
azed(y)
azed(y, preprocess = FALSE)
}
