% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sazed.R
\name{computeAcf}
\alias{computeAcf}
\title{Compute and shorten autocorrelation}
\usage{
computeAcf(y)
}
\arguments{
\item{y}{The input time series.}
}
\value{
The shortened autocorrelation
}
\description{
\code{computeAcf} computes the autocorrelation function of its argument and discards
the zero lag and all lags greater than 2/3 of the argument's length
}
\examples{
season_length <- 26
y <- sin(1:400*2*pi/season_length)
computeAcf(y)
}
