% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sazed.R
\name{sazed}
\alias{sazed}
\title{SAZED Ensemble (Optimum)}
\usage{
sazed(y)
}
\arguments{
\item{y}{The input time series.}
}
\value{
The season length of the input time series.
}
\description{
\code{sazed} estimates a time series' season length by combining 3 different estimates
computed on an input time series and its 10-fold self-composed autocorrelation.
}
\examples{
season_length <- 26
y <- sin(1:400*2*pi/season_length)
sazed(y)
}
