% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sazed.R
\docType{package}
\name{sazedR}
\alias{sazedR}
\alias{sazedR-package}
\title{sazedR: A package for for estimating the season length of a seasonal
time series.}
\description{
The sazedR package provides the main function to compute season length,
sazed, which is an ensemble of many season length estimation methods,
also included in this package.
}
